New IndividualGoalForm component for individual goals form on goals page
