+++ frontend/app/(root)/goals/page.tsx
@@ -46,6 +46,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card"
+import IndividualGoalForm from '@/components/IndividualGoalForm';
 
 
 export default function Goals() {
@@ -222,165 +223,22 @@ export default function Goals() {
             <button className='bg-bankGradient' style={styles.button}>Suggest Spending Changes</button>
           </Link>
         </div>
-        <Dialog open={open} onOpenChange={setOpen}>
-          <DialogTrigger asChild>
-              <h2
-                  className="text-14 font-semibold text-gray-600 view-all-btn cursor-pointer w-fit hover:scale-105 duration-300 ease-in-out"
-                  onClick={() => setOpen(true)}
-              >
-                  Add a New Goal
-              </h2>
-          </DialogTrigger>
-          <DialogContent className="bg-white">
-              <DialogHeader>
-                  <DialogTitle>Enter Individual Goal Details</DialogTitle>
-                  <DialogDescription>Add a new goal</DialogDescription>
-              </DialogHeader>
-              <div className="grid gap-4 py-4">
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="name" className="text-right">
-                          Name
-                      </Label>
-                      <Input
-                          id="name"
-                          value={name}
-                          onChange={(e) => setName(e.target.value)}
-                          className="col-span-3"
-                      />
-                  </div>
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="description" className="text-right">
-                          Description
-                      </Label>
-                      <Textarea
-                          id="description"
-                          value={description}
-                          onChange={(e) => setDescription(e.target.value)}
-                          className="col-span-3"
-                      />
-                  </div>
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="goalAmount" className="text-right">
-                          Goal Amount
-                      </Label>
-                      <Input
-                          id="goalAmount"
-                          value={goalAmount}
-                          onChange={(e) => setGoalAmount(e.target.value)}
-                          className="col-span-3"
-                      />
-                  </div>
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="amountContributed" className="text-right">
-                          Amount Contributed
-                      </Label>
-                      <Input
-                          id="amountContributed"
-                          value={amountContributed}
-                          onChange={(e) => setAmountContributed(e.target.value)}
-                          className="col-span-3"
-                      />
-                  </div>
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="targetDate" className="text-right">
-                          Target Date
-                      </Label>
-                      <Popover open={calendarOpen} onOpenChange={setCalendarOpen}>
-                        <PopoverTrigger asChild>
-                          <Button
-                            variant={"outline"}
-                            onClick={() => setCalendarOpen(true)}
-                            className={cn(
-                              "w-[280px] justify-start text-left font-normal",
-                              !targetDate && "text-muted-foreground"
-                            )}
-                          >
-                            <CalendarIcon className="mr-2 h-4 w-4" />
-                            {targetDate ? format(targetDate, "PPP") : <span>Pick a date</span>}
-                          </Button>
-                        </PopoverTrigger>
-                        <PopoverContent className="w-auto p-0 bg-white">
-                          <Calendar
-                            mode="single"
-                            selected={targetDate}
-                            onSelect={(selectedDate) => {
-                              setTargetDate(selectedDate);
-                              setCalendarOpen(false);
-                            }}
-                            initialFocus
-                          />
-                        </PopoverContent>
-                      </Popover>
-                  </div>
-                  <div className="grid grid-cols-4 items-center gap-4">
-                      <Label htmlFor="autoContribute" className="text-right">
-                          Auto Contribute
-                      </Label>
-                      <Checkbox id="autoContribute" onClick={() => setAutoContribute(!autoContribute)} />
-                  </div>
-                  {autoContribute && (
-                    <>
-                      <div className="grid grid-cols-4 items-center gap-4">
-                          <Label htmlFor="account" className="text-right">
-                              Account
-                          </Label>
-                          <Select onValueChange={(value) =>  {
-                            const selectedAccount = accounts.find((account: any) => account.accountId === value);
-                            if (selectedAccount) {
-                              setAccount(selectedAccount.accountId);
-                              setAccountName(selectedAccount.name);
-                            }
-                          }}>
-                              <SelectTrigger className="w-[180px] bg-white">
-                                  <SelectValue placeholder="Select" />
-                              </SelectTrigger>
-                              <SelectContent className="bg-white">
-                                  {accounts.map((account: any) => (
-                                      <SelectItem
-                                          key={account?.accountId}
-                                          value={account?.accountId}
-                                      >
-                                          {account?.name}
-                                      </SelectItem>
-                                  ))}
-                              </SelectContent>
-                          </Select>
-                      </div>
-                      <div className="grid grid-cols-4 items-center gap-4">
-                          <Label htmlFor="frequency" className="text-right">
-                              Frequency
-                          </Label>
-                          <Select onValueChange={(value) => setFrequency(value)}>
-                              <SelectTrigger className="w-[180px] bg-white">
-                                  <SelectValue placeholder="Select" />
-                              </SelectTrigger>
-                              <SelectContent className="bg-white">
-                                  <SelectItem value="DAILY">Daily</SelectItem>
-                                  <SelectItem value="WEEKLY">Weekly</SelectItem>
-                                  <SelectItem value="BIWEEKLY">Biweekly</SelectItem>
-                                  <SelectItem value="MONTHLY">Monthly</SelectItem>
-                              </SelectContent>
-                          </Select>
-                      </div>
-                      <div className="grid grid-cols-4 items-center gap-4">
-                          <Label htmlFor="frequencyNumber" className="text-right">
-                              Frequency Number
-                          </Label>
-                          <Input
-                              id="frequencyNumber"
-                              value={frequencyNumber}
-                              onChange={(e) => setFrequencyNumber(e.target.value)}
-                              className="col-span-3"
-                          />
-                      </div>
-                    </>
-                  )}
-              </div>
-              <DialogFooter>
-                  <Button onClick={() => addGoal()}>Add Goal</Button>
-              </DialogFooter>
-          </DialogContent>
-        </Dialog>
+        <IndividualGoalForm 
+          user={user} 
+          accounts={accounts}
+          onIndividualGoalAdded={async () => {
+            if(user?.userId) {
+                try {
+                  const response = await request('get', `/api/v1/individual-goals?userId=${user?.userId}`);
+                  console.log(response?.data);
+                  setIndividualGoals([...response?.data]);
+                } catch (error) {
+                  console.error(error);
+                }
+            }
+          }}
+        />
+
         
         <>
         {user?.role === "ADMIN" && (
